requirements.txt
# Required
schedule==1.2.2
tornado==6.3.3
marshmallow~=3.22
peewee>=3.14.4
peewee_migrate==1.13.0
psutil~=6.0.0
requests>=2.28.2
requests_toolbelt>=0.9.1
py-cpuinfo>=7.0.0
JSON-log-formatter~=0.5.2

# Optional requirements
watchdog>=2.1.1

# Required for CLI only
inquirer>=2.7.0

# Required for hosting the API docs
swagger-ui-py